Added KmlService and StringFormatterService
